+++ packages/cli/src/webpack/config.ts
@@ -21,7 +21,6 @@ import webpack, {
   WebpackPluginInstance,
 } from 'webpack';
 
-import { ImaConfigurationContext, ImaConfig } from '../types';
 import { GenerateRunnerPlugin } from './plugins/GenerateRunnerPlugin';
 import { ManifestPlugin } from './plugins/ManifestPlugin';
 import { createProgress } from './plugins/ProgressPlugin';
@@ -34,6 +33,7 @@ import {
   getCurrentCoreJsVersion,
   getLocaleEntryPoints,
 } from './utils';
+import { ImaConfigurationContext, ImaConfig } from '../types';
 
 /**
  * Creates Webpack configuration object based on input ConfigurationContext
@@ -59,6 +59,10 @@ export default async (
   const devServerConfig = createDevServerConfig({ imaConfig, ctx });
   const mode = ctx.environment === 'production' ? 'production' : 'development';
   const lessGlobalsPath = path.join(rootDir, 'app/less/globals.less');
+
+  // Bundle entries
+  const publicPathEntry = path.join(__dirname, './entries/publicPathEntry');
+  const appMainEntry = path.join(rootDir, 'app/main.js');
 
   // Define browserslist targets for current context
   const coreJsVersion = await getCurrentCoreJsVersion();
@@ -221,10 +225,11 @@ export default async (
     entry: {
       ...(isServer
         ? {
-            server: [path.join(rootDir, 'app/main.js')],
+            server: [publicPathEntry, appMainEntry],
           }
         : {
             [name]: [
+              publicPathEntry,
               useHMR &&
                 isDebug &&
                 `@ima/hmr-client?${new URLSearchParams({
@@ -237,7 +242,7 @@ export default async (
                   hostname: devServerConfig.hostname,
                   publicUrl: devServerConfig.publicUrl,
                 }).toString()}`,
-              path.join(rootDir, 'app/main.js'),
+              appMainEntry,
             ].filter(Boolean) as string[],
             ...createPolyfillEntry(ctx),
           }),