+++ packages/server/lib/factory/hooksFactory.js
@@ -101,7 +101,6 @@ module.exports = function hooksFactory({
     const isBadRequest = routeInfo && routeInfo.route.getName() === 'notFound';
 
     // TODO IMA@18 documentation badRequestConcurrency
-    //TODO IMA@18 update for better performance check
     return isBadRequest && _hasToServeStatic(event);
   }
 
@@ -235,12 +234,15 @@ module.exports = function hooksFactory({
   function useCreateContentVariablesHook() {
     emitter.on(Event.CreateContentVariables, async event => {
       if (!_isValidResponse(event)) {
-        return;
+        return event.result;
       }
 
-      event.context.response.contentVariables = createContentVariables({
-        ...event.context,
-      });
+      return {
+        ...event.result,
+        ...createContentVariables({
+          ...event.context,
+        }),
+      };
     });
   }
 
@@ -267,6 +269,10 @@ module.exports = function hooksFactory({
       }
 
       event = await emitter.emit(Event.CreateContentVariables, event);
+      event.context.response.contentVariables = {
+        ...event.context.response.contentVariables,
+        ...event.result,
+      };
 
       event.context.response.content = processContent({
         ...event.context,